+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file. The format 
 
 ## [Unreleased]
 
+### Added
+- TeX-style rendering option for Inline Numerals results in Live Preview and Reading mode. (Closes #161)
+
 ## [1.10.2] - 2026-05-19
 
 ### Changed